+++ ldap4j-tests/src/test/java/hu/gds/ldap4j/ldap/UnboundidDSTest.java
@@ -13,6 +13,7 @@ import hu.gds.ldap4j.ldap.extension.FeatureDiscovery;
 import hu.gds.ldap4j.ldap.extension.ManageDsaIt;
 import hu.gds.ldap4j.ldap.extension.ModifyIncrement;
 import hu.gds.ldap4j.ldap.extension.PasswordModify;
+import hu.gds.ldap4j.ldap.extension.ReadEntryControls;
 import hu.gds.ldap4j.ldap.extension.ServerSideSorting;
 import hu.gds.ldap4j.ldap.extension.SimplePagedResults;
 import hu.gds.ldap4j.ldap.extension.WhoAmI;
@@ -860,6 +861,48 @@ public class UnboundidDSTest {
                                             .composeIgnoreResult(()->Lava.VOID);
                                 }
                             }));
+        }
+    }
+
+    @Test
+    public void testReadEntryControls() throws Throwable {
+        try (TestContext<LdapTestParameters> context=TestContext.create(TEST_PARAMETERS);
+             UnboundidDirectoryServer ldapServer=new UnboundidDirectoryServer(
+                     false, TEST_PARAMETERS.serverPortClearText, TEST_PARAMETERS.serverPortTls)) {
+            ldapServer.start();
+            context.get(
+                    Closeable.withCloseable(
+                            ()->context.parameters().connectionFactory(
+                                    context, ldapServer, UnboundidDirectoryServer.adminBind()),
+                            (connection)->connection.writeRequestReadResponseChecked(
+                                            new ModifyRequest(
+                                                    List.of(
+                                                            new ModifyRequest.Change(
+                                                                    new PartialAttribute(
+                                                                            "objectClass",
+                                                                            List.of("extensibleObject")),
+                                                                    ModifyRequest.OPERATION_ADD)),
+                                                    "ou=groups,ou=test,dc=ldap4j,dc=gds,dc=hu")
+                                                    .controls(List.of(
+                                                            ReadEntryControls.postRequest(List.of("objectClass")),
+                                                            ReadEntryControls.preRequest(List.of("objectClass")))))
+                                    .compose((response)->{
+                                        @NotNull SearchResult.Entry postEntry
+                                                =ReadEntryControls.postResponseCheck(response.controls());
+                                        assertEquals(1, postEntry.attributes().size());
+                                        assertEquals("objectClass", postEntry.attributes().get(0).type());
+                                        assertEquals(
+                                                List.of("top", "organizationalUnit", "extensibleObject"),
+                                                postEntry.attributes().get(0).values());
+                                        @NotNull SearchResult.Entry preEntry
+                                                =ReadEntryControls.preResponseCheck(response.controls());
+                                        assertEquals(1, preEntry.attributes().size());
+                                        assertEquals("objectClass", preEntry.attributes().get(0).type());
+                                        assertEquals(
+                                                List.of("top", "organizationalUnit"),
+                                                preEntry.attributes().get(0).values());
+                                        return Lava.VOID;
+                                    })));
         }
     }